Changes for registration
