--- app/src/main/res/layout/activity_trucker_registration_t.xml
@@ -1,63 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".registration.TruckerRegistrationT">
-
-    <EditText
-        android:id="@+id/truck_name"
-        android:layout_width="match_parent"
-        android:layout_marginTop="100dp"
-        android:layout_height="40dp"
-        android:layout_marginLeft="20dp"
-        android:layout_marginRight="20dp"
-        android:hint="TRUCK NAME"
-        app:layout_constraintBottom_toTopOf="@+id/truck_type"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
-    <EditText
-        android:id="@+id/truck_type"
-        android:layout_width="match_parent"
-        android:layout_height="40dp"
-        android:layout_margin="20dp"
-        android:hint="TRUCK TYPE"
-        app:layout_constraintBottom_toTopOf="@+id/max_length"
-        app:layout_constraintEnd_toEndOf="@+id/max_length"
-        app:layout_constraintStart_toStartOf="@+id/max_length"
-        app:layout_constraintTop_toBottomOf="@+id/truck_name" />
-
-    <EditText
-        android:id="@+id/max_length"
-        android:layout_width="match_parent"
-        android:layout_height="40dp"
-        android:layout_margin="20dp"
-        android:hint="max length"
-        app:layout_constraintBottom_toTopOf="@+id/max_wt"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/truck_type" />
-
-    <EditText
-        android:id="@+id/max_wt"
-        android:layout_width="match_parent"
-        android:layout_height="40dp"
-        android:layout_margin="20dp"
-        android:hint="WT#"
-        app:layout_constraintBottom_toTopOf="@+id/registerTruck"
-        app:layout_constraintEnd_toEndOf="@+id/truck_name"
-        app:layout_constraintStart_toStartOf="@+id/max_length"
-        app:layout_constraintTop_toBottomOf="@+id/max_length" />
-
-    <Button
-        android:id="@+id/registerTruck"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_margin="20dp"
-        android:text="register"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/max_wt" />
-</androidx.constraintlayout.widget.ConstraintLayout>
