+++ app/src/main/res/layout/activity_trucker_registration_two.xml
@@ -111,7 +111,7 @@
         app:layout_constraintVertical_bias="0.0" />
 
     <Button
-        android:id="@+id/postLoadButton"
+        android:id="@+id/finalRegButton"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:layout_margin="20dp"
